# remove default host attribute in router